move DatasetConfig
