+++ nexus/reconfigurator/planning/src/system.rs
@@ -56,6 +56,7 @@ use omicron_uuid_kinds::MupdateOverrideUuid;
 use omicron_uuid_kinds::OmicronZoneUuid;
 use omicron_uuid_kinds::SledUuid;
 use omicron_uuid_kinds::ZpoolUuid;
+use sled_agent_types::disk::DatasetConfig;
 use sled_agent_types::disk::DiskIdentity;
 use sled_agent_types::disk::DiskVariant;
 use sled_agent_types::disk::M2Slot;
@@ -1736,10 +1737,7 @@ impl Sled {
     ///
     /// The inventory values for "available space" and "used space" are
     /// made up, since this is a synthetic dataset.
-    pub fn add_synthetic_dataset(
-        &mut self,
-        config: omicron_common::disk::DatasetConfig,
-    ) {
+    pub fn add_synthetic_dataset(&mut self, config: DatasetConfig) {
         self.inventory_sled_agent.datasets.push(InventoryDataset {
             id: Some(config.id),
             name: config.name.full_name(),
