+++ nexus/src/app/background/tasks/support_bundle_collector.rs
@@ -727,7 +727,6 @@ mod test {
     use nexus_types::support_bundle::BundleDataSelection;
     use omicron_common::api::external::ByteCount;
     use omicron_common::api::internal::shared::DatasetKind;
-    use omicron_common::disk::DatasetConfig;
     use omicron_common::disk::DatasetName;
     use omicron_common::disk::SharedDatasetConfig;
     use omicron_common::zpool_name::ZpoolName;
@@ -736,6 +735,7 @@ mod test {
         BlueprintUuid, DatasetUuid, EreporterRestartUuid, OmicronZoneUuid,
         PhysicalDiskUuid, RackUuid, SledUuid,
     };
+    use sled_agent_types::disk::DatasetConfig;
     use sled_agent_types::inventory::ZpoolHealth;
     use std::num::NonZeroU64;
     use support_bundle_collection::perfetto;
