+++ sled-agent/src/support_bundle/storage.rs
@@ -17,7 +17,6 @@ use illumos_utils::zfs::DatasetProperties;
 use illumos_utils::zfs::DestroyDatasetError;
 use illumos_utils::zfs::DestroyDatasetErrorVariant;
 use omicron_common::disk::CompressionAlgorithm;
-use omicron_common::disk::DatasetConfig;
 use omicron_common::disk::DatasetName;
 use omicron_common::disk::SharedDatasetConfig;
 use omicron_uuid_kinds::DatasetUuid;
@@ -33,6 +32,7 @@ use sled_agent_config_reconciler::NestedDatasetDestroyError;
 use sled_agent_config_reconciler::NestedDatasetEnsureError;
 use sled_agent_config_reconciler::NestedDatasetListError;
 use sled_agent_config_reconciler::NestedDatasetMountError;
+use sled_agent_types::disk::DatasetConfig;
 use sled_agent_types::support_bundle::BUNDLE_FILE_NAME;
 use sled_agent_types::support_bundle::BUNDLE_TMP_FILE_NAME;
 use sled_agent_types::support_bundle::NESTED_DATASET_NOT_FOUND;
@@ -1117,13 +1117,13 @@ mod tests {
         ACCEPT_RANGES, CONTENT_LENGTH, CONTENT_RANGE, CONTENT_TYPE,
     };
     use iddqd::id_ord_map;
-    use omicron_common::disk::DatasetConfig;
     use omicron_common::disk::DatasetKind;
     use omicron_common::disk::DatasetName;
     use omicron_common::zpool_name::ZpoolName;
     use omicron_test_utils::dev::test_setup_log;
     use omicron_uuid_kinds::PhysicalDiskUuid;
     use sha2::Sha256;
+    use sled_agent_types::disk::DatasetConfig;
     use sled_agent_types::inventory::ZpoolHealth;
     use uuid::Uuid;
     use zip::ZipWriter;
