+++ sled-agent/types/versions/src/add_dual_stack_external_ip_config/inventory.rs
@@ -13,7 +13,6 @@ use iddqd::id_upcast;
 use omicron_common::{
     address::{Ip, NUM_SOURCE_NAT_PORTS},
     api::external::{ByteCount, Generation},
-    disk::DatasetConfig,
     zpool_name::ZpoolName,
 };
 use omicron_ledger::Ledgerable;
@@ -24,6 +23,7 @@ use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 use crate::impls::inventory::SourceNatConfigError;
+use crate::v1::disk::DatasetConfig;
 use crate::v1::disk::OmicronPhysicalDiskConfig;
 use crate::v1::inventory::Baseboard;
 use crate::v1::inventory::{
