+++ sled-agent/types/versions/src/add_nexus_lockstep_port_to_inventory/inventory.rs
@@ -12,7 +12,6 @@ use iddqd::IdOrdMap;
 use iddqd::id_upcast;
 use omicron_common::address::NEXUS_LOCKSTEP_PORT;
 use omicron_common::api::external::{ByteCount, Generation};
-use omicron_common::disk::DatasetConfig;
 use omicron_common::zpool_name::ZpoolName;
 use omicron_ledger::Ledgerable;
 use omicron_uuid_kinds::{DatasetUuid, MupdateOverrideUuid, OmicronZoneUuid};
@@ -21,6 +20,7 @@ use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 use crate::v1;
+use crate::v1::disk::DatasetConfig;
 use crate::v1::disk::OmicronPhysicalDiskConfig;
 use crate::v1::inventory::Baseboard;
 use crate::v1::inventory::{
